replace all input search box with typeahed select for all filters
