+++ console/console-init/ui/src/Pages/AddressSpaceDetail/ConnectionList/ConnectionListFilter.tsx
@@ -1,0 +1,396 @@
+import * as React from "react";
+import {
+  DataToolbarChip,
+  DataToolbarGroup,
+  DataToolbarFilter,
+  DataToolbarItem,
+  DataToolbarToggleGroup,
+  DataToolbar,
+  DataToolbarContent
+} from "@patternfly/react-core/dist/js/experimental";
+import {
+  Dropdown,
+  DropdownToggle,
+  DropdownItem,
+  InputGroup,
+  TextInput,
+  Button,
+  ButtonVariant,
+  Badge,
+  SelectOption,
+  SelectOptionObject,
+  Select,
+  SelectVariant
+} from "@patternfly/react-core";
+import { FilterIcon, SearchIcon } from "@patternfly/react-icons";
+import useWindowDimensions from "src/Components/Common/WindowDimension";
+import { ISortBy } from "@patternfly/react-table";
+import { SortForMobileView } from "src/Components/Common/SortForMobileView";
+
+interface IConnectionListFilterProps {
+  filterValue?: string | null;
+  setFilterValue: (value: string) => void;
+  hostnames?: Array<string>;
+  setHostnames: (value: Array<string>) => void;
+  containerIds?: Array<string>;
+  setContainerIds: (value: Array<string>) => void;
+  sortValue?: ISortBy;
+  setSortValue: (value: ISortBy) => void;
+  totalConnections: number;
+}
+export const ConnectionListFilter: React.FunctionComponent<IConnectionListFilterProps> = ({
+  filterValue,
+  setFilterValue,
+  hostnames,
+  setHostnames,
+  containerIds,
+  setContainerIds,
+  totalConnections,
+  sortValue,
+  setSortValue
+}) => {
+  const { width } = useWindowDimensions();
+  const [inputValue, setInputValue] = React.useState<string | null>();
+  const [filterIsExpanded, setFilterIsExpanded] = React.useState(false);
+  const onFilterSelect = (event: any) => {
+    setFilterValue(event.target.value);
+    setFilterIsExpanded(!filterIsExpanded);
+  };
+  const [isSelectHostnameExpanded, setIsSelectHostnameExpanded] = React.useState<
+    boolean
+  >(false);
+  const [isSelectContainerExpanded, setIsSelectContainerExpanded] = React.useState<
+    boolean
+  >(false);
+  const [hostnameSelected, setHostnameSelected] = React.useState<string>();
+  const [containerSelected, setContainerSelected] = React.useState<string>();
+  const [hostnameOptions, setHostnameOptions] = React.useState<Array<string>>();
+  const [containerOptions, setContainerOptions] = React.useState<Array<string>>();
+
+  const filterMenuItems = [
+    { key: "filterHostName", value: "Hostname" },
+    { key: "filterContainer", value: "Container" }
+  ];
+
+  const sortMenuItems = [
+    { key: "hostname", value: "Hostname", index: 0 },
+    { key: "containerId", value: "Container ID", index: 1 },
+    { key: "protocol", value: "Protocol", index: 2 },
+    { key: "messageIn", value: "Messages In", index: 3 },
+    { key: "messageOut", value: "Messages Out", index: 4 },
+    { key: "sender", value: "Senders", index: 5 },
+    { key: "receiver", value: "Receivers", index: 6 }
+  ];
+
+  const onInputChange = (newValue: string) => {
+    setInputValue(newValue);
+  };
+  const onClickSearchIcon = (event: any) => {
+    if (filterValue)
+      if (filterValue === "Container") {
+        if (containerSelected && containerSelected.trim() !== "" && containerIds) {
+          if (containerIds.indexOf(containerSelected.trim()) < 0) {
+            setContainerIds([...containerIds, containerSelected.trim()]);
+            setContainerSelected(undefined)
+          }
+          setInputValue(null);
+        }
+      } else if (filterValue === "Hostname") {
+        if (hostnameSelected && hostnameSelected.trim() !== "" && hostnames) {
+          if (hostnames.indexOf(hostnameSelected) < 0) {
+            setHostnames([...hostnames, hostnameSelected]);
+          }
+          setHostnameSelected(undefined);
+        }
+      }
+  };
+
+
+  const onHostnameSelectToggle = () => {
+    setIsSelectHostnameExpanded(!isSelectHostnameExpanded);
+  };
+
+  const onContainerSelectToggle = () => {
+    setIsSelectContainerExpanded(!isSelectContainerExpanded);
+  };
+
+  const onChangeHostnameData = async (value: string) => {
+    setHostnameOptions(undefined);
+    // const response = await client.query<IConnectionLinksNameSearchResponse>({
+    //   query: RETURN_ALL_CONNECTION_LINKS_FOR_NAME_SEARCH(
+    //     connectionName,
+    //     addressSpaceName,
+    //     namespace,
+    //     value.trim()
+    //   )
+    // });
+    // if (
+    //   response &&
+    //   response.data &&
+    //   response.data.connections &&
+    //   response.data.connections.Connections &&
+    //   response.data.connections.Connections.length > 0 &&
+    //   response.data.connections.Connections[0].Links &&
+    //   response.data.connections.Connections[0].Links.Links &&
+    //   response.data.connections.Connections[0].Links.Links.length > 0
+    // ) {
+    //   const obtainedList = response.data.connections.Connections[0].Links.Links.map(
+    //     (link: any) => {
+    //       console.log("Name", link);
+    //       return link.ObjectMeta.Name;
+    //     }
+    //   );
+    //   setNameOptions(obtainedList);
+    // }
+  };
+
+  const onHostnameSelectFilterChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    onChangeHostnameData(e.target.value);
+    const options: React.ReactElement[] = hostnameOptions
+      ? hostnameOptions.map((option, index) => (
+          <SelectOption key={index} value={option} />
+        ))
+      : [];
+    return options;
+  };
+
+  const onChangeContainerData = async (value: string) => {
+    setContainerOptions(undefined);
+    // const response = await client.query<IConnectionLinksAddressSearchResponse>({
+    //   query: RETURN_ALL_CONNECTION_LINKS_FOR_ADDRESS_SEARCH(
+    //     connectionName,
+    //     addressSpaceName,
+    //     namespace,
+    //     value.trim()
+    //   )
+    // });
+    // if (
+    //   response &&
+    //   response.data &&
+    //   response.data.connections &&
+    //   response.data.connections.Connections &&
+    //   response.data.connections.Connections.length > 0 &&
+    //   response.data.connections.Connections[0].Links &&
+    //   response.data.connections.Connections[0].Links.Links &&
+    //   response.data.connections.Connections[0].Links.Links.length > 0
+    // ) {
+    //   const obtainedList = response.data.connections.Connections[0].Links.Links.map(
+    //     (link: any) => {
+    //       return link.Spec.Address;
+    //     }
+    //   );
+    //   setAddressOptions(obtainedList);
+    // }
+  };
+  const onContainerSelectFilterChange = (
+    e: React.ChangeEvent<HTMLInputElement>
+  ) => {
+    onChangeContainerData(e.target.value);
+    const options: React.ReactElement[] = containerOptions
+      ? containerOptions.map((option, index) => (
+          <SelectOption key={index} value={option} />
+        ))
+      : [];
+    return options;
+  };
+
+  const onHostnameSelect = (event: any, selection: string | SelectOptionObject) => {
+    setHostnameSelected(selection.toString());
+    setIsSelectHostnameExpanded(false);
+  };
+
+  const onContainerSelect = (
+    event: any,
+    selection: string | SelectOptionObject
+  ) => {
+    setContainerSelected(selection.toString());
+    setIsSelectContainerExpanded(false);
+  };
+
+  const onDelete = (
+    type: string | DataToolbarChip,
+    id: string | DataToolbarChip
+  ) => {
+    let index;
+
+    switch (type) {
+      case "Hostname":
+        if (hostnames && id) {
+          index = hostnames.indexOf(id.toString());
+          if (index >= 0) hostnames.splice(index, 1);
+          setHostnames([...hostnames]);
+        }
+        break;
+      case "Container":
+        if (containerIds && id) {
+          const containers = containerIds;
+          index = containerIds.indexOf(id.toString());
+          if (index >= 0) containers.splice(index, 1);
+          setContainerIds([...containers]);
+        }
+        break;
+    }
+  };
+  const clearAllFilters = () => {
+    setHostnames([]);
+    setContainerIds([]);
+  };
+
+  const checkIsFilterApplied = () => {
+    if (
+      (containerIds && containerIds.length > 0) ||
+      (hostnames && hostnames.length > 0)
+    ) {
+      return true;
+    }
+    return false;
+  };
+
+  const toggleGroupItems = (
+    <>
+      <DataToolbarGroup variant="filter-group">
+        <DataToolbarFilter categoryName="Filter">
+          <Dropdown
+            id="cl-filter-dropdown"
+            position="left"
+            onSelect={onFilterSelect}
+            isOpen={filterIsExpanded}
+            toggle={
+              <DropdownToggle onToggle={setFilterIsExpanded}>
+                <FilterIcon />
+                &nbsp;
+                {filterValue && filterValue.trim() !== ""
+                  ? filterValue
+                  : "Filter"}
+              </DropdownToggle>
+            }
+            dropdownItems={filterMenuItems.map(option => (
+              <DropdownItem
+                key={option.key}
+                value={option.value}
+                itemID={option.key}
+                component={"button"}>
+                {option.value}
+              </DropdownItem>
+            ))}
+          />
+        </DataToolbarFilter>
+        <DataToolbarItem>
+          <DataToolbarFilter
+            chips={hostnames}
+            deleteChip={onDelete}
+            categoryName="Hostname">
+            {filterValue && filterValue === "Hostname" && (
+              <InputGroup>
+                <Select
+                    variant={SelectVariant.typeahead}
+                    aria-label="Select a hostname"
+                    onToggle={onHostnameSelectToggle}
+                    onSelect={onHostnameSelect}
+                    onClear={() => {
+                      setHostnameSelected(undefined);
+                      setIsSelectHostnameExpanded(false);
+                    }}
+                    selections={hostnameSelected}
+                    onFilter={onHostnameSelectFilterChange}
+                    isExpanded={isSelectHostnameExpanded}
+                    ariaLabelledBy={"typeahead-select-id"}
+                    placeholderText="Select Hostname"
+                    isDisabled={false}
+                    isCreatable={false}>
+                    {hostnameOptions &&
+                      hostnameOptions.map((option, index) => (
+                        <SelectOption key={index} value={option} />
+                      ))}
+                    {/* {} */}
+                  </Select>
+                <Button
+                  id="cl-filter-search-btn"
+                  variant={ButtonVariant.control}
+                  aria-label="search button for search input"
+                  onClick={onClickSearchIcon}>
+                  <SearchIcon />
+                </Button>
+              </InputGroup>
+            )}
+          </DataToolbarFilter>
+        </DataToolbarItem>
+        <DataToolbarItem>
+          <DataToolbarFilter
+            chips={containerIds}
+            deleteChip={onDelete}
+            categoryName="Container">
+            {filterValue && filterValue === "Container" && (
+              <InputGroup>
+                <Select
+                    variant={SelectVariant.typeahead}
+                    aria-label="Select a Address"
+                    onToggle={onContainerSelectToggle}
+                    onSelect={onContainerSelect}
+                    onClear={() => {
+                      setContainerSelected(undefined);
+                      setIsSelectContainerExpanded(false);
+                    }}
+                    selections={containerSelected}
+                    onFilter={onContainerSelectFilterChange}
+                    isExpanded={isSelectContainerExpanded}
+                    ariaLabelledBy={"typeahead-select-id"}
+                    placeholderText="Select Container"
+                    isDisabled={false}
+                    isCreatable={false}>
+                    {containerOptions &&
+                      containerOptions.map((option, index) => (
+                        <SelectOption key={index} value={option} />
+                      ))}
+                    {/* {} */}
+                  </Select>
+                <Button
+                  variant={ButtonVariant.control}
+                  aria-label="search button for search input"
+                  onClick={onClickSearchIcon}>
+                  <SearchIcon />
+                </Button>
+              </InputGroup>
+            )}
+          </DataToolbarFilter>
+        </DataToolbarItem>
+      </DataToolbarGroup>
+    </>
+  );
+  const toolbarItems = (
+    <>
+      <DataToolbarToggleGroup
+        toggleIcon={
+          <>
+            <FilterIcon />
+            {checkIsFilterApplied() && (
+              <Badge key={1} isRead>
+                {totalConnections}
+              </Badge>
+            )}
+          </>
+        }
+        breakpoint="xl">
+        {toggleGroupItems}
+      </DataToolbarToggleGroup>
+      {width < 769 && (
+        <SortForMobileView
+          sortMenu={sortMenuItems}
+          sortValue={sortValue}
+          setSortValue={setSortValue}
+        />
+      )}
+    </>
+  );
+  return (
+    <>
+      <DataToolbar
+        id="data-toolbar-with-filter"
+        className="pf-m-toggle-group-container"
+        collapseListedFiltersBreakpoint="md"
+        clearAllFilters={clearAllFilters}>
+        <DataToolbarContent>{toolbarItems}</DataToolbarContent>
+      </DataToolbar>
+    </>
+  );
+};
