+++ console/console-init/ui/src/stories/AddressDetailHeader.stories.tsx
@@ -3,7 +3,7 @@ import { MemoryRouter } from "react-router";
 import { text, number, select } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
 import { AddressDetailHeader } from "../Components/AddressDetail/AddressDetailHeader";
-import { AddressLinksFilter } from "src/Components/AddressDetail/AddressLinksFilter";
+import { AddressLinksFilter } from "src/Pages/AddressDetail/AddressLinksFilter";
 
 export default {
   title: "Address Detail"
